+++ includes/post-types-admin.php
@@ -1722,11 +1722,12 @@ function radio_station_show_info_metabox() {
 
 	// --- get show meta ---
 	// 2.3.2: added show download disable switch
-	$file = get_post_meta( $post->ID, 'show_file', true );
-	$download = get_post_meta( $post->ID, 'show_download', true );
-	$email = get_post_meta( $post->ID, 'show_email', true );
 	$active = get_post_meta( $post->ID, 'show_active', true );
 	$link = get_post_meta( $post->ID, 'show_link', true );
+	$email = get_post_meta( $post->ID, 'show_email', true );
+	$phone = get_post_meta( $post->ID, 'show_phone', true );
+	$file = get_post_meta( $post->ID, 'show_file', true );
+	$download = get_post_meta( $post->ID, 'show_download', true );
 	$patreon_id = get_post_meta( $post->ID, 'show_patreon', true );
 
 	// added max-width to prevent metabox overflows
@@ -1741,8 +1742,13 @@ function radio_station_show_info_metabox() {
 		echo '<p><div style="width:120px; display:inline-block;"><label>' . esc_html( __( 'Website Link', 'radio-station' ) ) . ':</label></div>
 		<input type="text" name="show_link" size="100" style="max-width:80%;" value="' . esc_url( $link ) . '" /></p>';
 
-		echo '<p><div style="width:120px; display:inline-block;"><label>' . esc_html( __( 'DJ / Host Email', 'radio-station' ) ) . ':</label></div>
+		// 2.3.3.6: change text string from DJ / Host email (as maybe multiple hosts)
+		echo '<p><div style="width:120px; display:inline-block;"><label>' . esc_html( __( 'Show Email', 'radio-station' ) ) . ':</label></div>
 		<input type="text" name="show_email" size="100" style="max-width:80%;" value="' . esc_attr( $email ) . '" /></p>';
+
+		// 2.3.3.6: added Show phone number input field
+		echo '<p><div style="width:120px; display:inline-block;"><label>' . esc_html( __( 'Show Phone', 'radio-station' ) ) . ':</label></div>
+		<input type="text" name="show_phone" size="100" style="max-width:80%;" value="' . esc_attr( $phone ) . '" /></p>';
 
 		echo '<p><div style="width:120px; display:inline-block;"><label>' . esc_html( __( 'Latest Audio File', 'radio-station' ) ) . ':</label></div>
 		<input type="text" name="show_file" size="100" style="max-width:80%;" value="' . esc_attr( $file ) . '" /></p>';
@@ -3135,6 +3141,18 @@ function radio_station_show_save_data( $post_id ) {
 		$link = filter_var( trim( $_POST['show_link'] ), FILTER_SANITIZE_URL );
 		$patreon_id = sanitize_title( $_POST['show_patreon'] );
 
+		// 2.3.3.6: added phone number with character filter validation
+		$phone = trim( $_POST['show_phone'] );
+		if ( strlen( $phone ) > 0 ) {
+			$phone = str_split( $phone, 1 );
+			$phone = preg_filter( '/^[0-9+\(\)#\.\s\-]+$/', '$0', $phone );
+			if ( count( $phone ) > 0 ) {
+				$phone = implode( '', $phone );
+			} else {
+				$phone = '';
+			}
+		}
+
 		// 2.2.8: removed strict in_array checking
 		// 2.3.2: fix for unchecked boxes index warning
 		$active = $download = '';
@@ -3155,15 +3173,18 @@ function radio_station_show_save_data( $post_id ) {
 		// --- get existing values and check if changed ---
 		// 2.3.0: added check against previous values
 		// 2.3.2: added download disable switch
+		// 2.3.3.6: added phone number field saving
 		$prev_file = get_post_meta( $post_id, 'show_file', true );
 		$prev_download = get_post_meta( $post_id, 'show_download', true );
 		$prev_email = get_post_meta( $post_id, 'show_email', true );
+		$prev_phone = get_post_meta( $post_id, 'show_phone', true );
 		$prev_active = get_post_meta( $post_id, 'show_active', true );
 		$prev_link = get_post_meta( $post_id, 'show_link', true );
 		$prev_patreon_id = get_post_meta( $post_id, 'show_patreon', true );
-		if ( ( $prev_file != $file ) || ( $prev_email != $email )
-		     || ( $prev_active != $active ) || ( $prev_link != $link )
-		     || ( $prev_download != $download ) || ( $prev_patreon_id != $patreon_id ) ) {
+		if ( ( $prev_active != $active ) || ( $prev_link != $link )
+		     || ( $prev_email != $email ) || ( $prev_phone != $phone )
+		     || ( $prev_file != $file ) || ( $prev_download != $download )
+		     || ( $prev_patreon_id != $patreon_id ) ) {
 			$show_meta_changed = true;
 		}
 
@@ -3172,6 +3193,7 @@ function radio_station_show_save_data( $post_id ) {
 		update_post_meta( $post_id, 'show_file', $file );
 		update_post_meta( $post_id, 'show_download', $download );
 		update_post_meta( $post_id, 'show_email', $email );
+		update_post_meta( $post_id, 'show_phone', $phone );
 		update_post_meta( $post_id, 'show_active', $active );
 		update_post_meta( $post_id, 'show_link', $link );
 		update_post_meta( $post_id, 'show_patreon', $patreon_id );
@@ -3383,8 +3405,9 @@ function radio_station_show_save_data( $post_id ) {
 	if ( defined( 'DOING_AJAX' ) && DOING_AJAX ) {
 		if ( isset( $_POST['action'] ) && ( 'radio_station_show_save_shifts' == $_POST['action'] ) ) {
 
-			print_r( $_POST['show_sched'] );
-			print_r( $new_shifts );
+			// --- (hidden) debug information ---
+			echo "Posted Shifsts: " . print_r( $_POST['show_sched'], true ) . PHP_EOL;
+			echo "New Shifts: " . print_r( $new_shifts, true ) . PHP_EOL;
 
 			// --- display shifts saved message ---
 			$show_shifts_nonce = wp_create_nonce( 'radio-station' );